feat: polish globe visuals
